Update our own Topology class
